Added mobile version cards
