+++ package.json
@@ -1,180 +1,180 @@
 {
-  
+  
   "name": "carzilla",
-  
+  
   "version": "0.1.0",
-  
+  
   "private": true,
-  
+  
   "dependencies": {
-  
-    
+  
+    
   "@emailjs/browser": "^4.3.3",
-  
-    
+  
+    
   "@emotion/react": "^11.11.4",
-  
-    
+  
+    
   "@emotion/styled": "^11.11.0",
-  
-    
+  
+    
   "@fortawesome/fontawesome-svg-core": "^6.5.1",
-  
-    
+  
+    
   "@fortawesome/free-brands-svg-icons": "^6.5.1",
-  
-    
+  
+    
   "@fortawesome/free-regular-svg-icons": "^6.5.1",
-  
-    
+  
+    
   "@fortawesome/free-solid-svg-icons": "^6.5.1",
-  
-    
+  
+    
   "@fortawesome/react-fontawesome": "^0.2.0",
-  
-    
+  
+    
   "@mui/icons-material": "^5.15.13",
-  
-    
+  
+    
   "@mui/material": "^5.15.13",
-  
-    
+  
+    
   "@mui/styled-engine-sc": "^6.0.0-alpha.17",
-  
-    
+  
+    
   "@testing-library/jest-dom": "^5.17.0",
-  
-    
+  
+    
   "@testing-library/react": "^13.4.0",
-  
-    
+  
+    
   "@testing-library/user-event": "^13.5.0",
-  
-    
+  
+    
   "aos": "^2.3.4",
-  
-    
+  
+    
   "axios": "^1.7.2",
-  
-    
+  
+    
   "bootstrap": "^5.3.3",
-  
-    
+  
+    
   "react": "^18.2.0",
-  
-    
+  
+    
   "react-bootstrap": "^2.10.1",
-  
-    
+  
+    
   "react-dom": "^18.2.0",
-  
-    
+  
+    
   "react-multi-carousel": "^2.8.5",
-  
-    
+  
+    
   "react-router-dom": "^6.22.3",
-  
-    
+  
+    
   "react-scripts": "5.0.1",
-  
-    
+  
+    
   "react-slick": "^0.30.2",
-  
-    
+  
+    
   "slick-carousel": "^1.8.1",
-  
-    
+  
+    
   "styled-components": "^6.1.8",
-  
-    
+  
+    
   "web-vitals": "^2.1.4"
-  
+  
   },
-  
+  
   "scripts": {
-  
-    
+  
+    
   "start": "react-scripts start",
-  
-    
+  
+    
   "predeploy": "npm run build",
-  
-    
+  
+    
   "deploy": "gh-pages -d build",
-  
-    
+  
+    
   "build": "react-scripts build",
-  
-    
+  
+    
   "test": "react-scripts test",
-  
-    
+  
+    
   "eject": "react-scripts eject"
-  
+  
   },
-  
+  
   "eslintConfig": {
-  
-    
+  
+    
   "extends": [
-  
-    
-    
+  
+    
+    
   "react-app",
-  
-    
-    
+  
+    
+    
   "react-app/jest"
-  
-    
+  
+    
   ]
-  
+  
   },
-  
+  
   "browserslist": {
-  
-    
+  
+    
   "production": [
-  
-    
-    
+  
+    
+    
   ">0.2%",
-  
-    
-    
+  
+    
+    
   "not dead",
-  
-    
-    
+  
+    
+    
   "not op_mini all"
-  
-    
+  
+    
   ],
-  
-    
+  
+    
   "development": [
-  
-    
-    
+  
+    
+    
   "last 1 chrome version",
-  
-    
-    
+  
+    
+    
   "last 1 firefox version",
-  
-    
-    
+  
+    
+    
   "last 1 safari version"
-  
-    
+  
+    
   ]
-  
+  
   },
-  
+  
   "devDependencies": {
-  
-    
+  
+    
   "gh-pages": "^6.1.1"
-  
+  
   }
 }
