+++ src/index.css
@@ -1621,3 +1621,101 @@ button{
   font-weight: 700;
   text-align: center;
 }
+
+
+
+
+
+
+
+.services-container {
+  width: 100%;
+  max-width: 900px;
+  margin: 0 auto;
+  padding: 20px 40px; 
+  background: #F7F7F7;
+}
+
+.cards-grid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 20px 24px;
+  max-width: 600px;
+  margin-bottom: 40px;
+}
+.cards-grid:hover{
+  cursor: pointer;
+}
+
+.card-mobile {
+  position: relative;
+  border-radius: 16px;
+  height: 140px; 
+  width: 100%;
+  aspect-ratio: 4/3; 
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 16px;
+}
+
+.card-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+}
+
+.card-mobile h3 {
+  color: white;
+  font-size: 20px;
+  font-weight: 600;
+  margin: 0;
+  margin-bottom: 2px;
+}
+
+.card-mobile p {
+  color: white;
+  font-size: 12px;
+  margin: 0;
+  opacity: 0.9;
+}
+
+.car-image-wrapper {
+  margin-top: 16px;
+}
+
+.car-image {
+  width: 108px;
+  height: 37px;
+  object-fit: contain;
+}
+
+/* Media Queries */
+@media screen and (max-width: 768px) {
+  .services-container {
+      padding: 16px;
+  }
+
+  .cards-grid {
+      gap: 16px;
+  }
+
+  .card-mobile {
+      height: 130px;
+  }
+
+  .card-mobile h3 {
+      font-size: 18px;
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .cards-grid {
+      grid-template-columns: 1fr;
+  }
+
+  .card-mobile {
+      height: 120px;
+  }
+}
